Make Snappy optional in JVM mode
Fix https://github.com/quarkusio/quarkus/issues/34670.
